consumer/requirements.txt
kafka-python==2.1.1
lz4==4.3.3
